+++ src/main/java/com/ll/commars/domain/auth/service/AuthService.java
@@ -6,20 +6,15 @@ import java.time.Duration;
 import java.time.Instant;
 
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.ll.commars.domain.auth.dto.SignInRequest;
-import com.ll.commars.domain.auth.dto.SignInResponse;
 import com.ll.commars.domain.auth.dto.TokenReissueResponse;
 import com.ll.commars.domain.auth.token.TokenProvider;
 import com.ll.commars.domain.auth.token.entity.AccessToken;
 import com.ll.commars.domain.auth.token.entity.JwtClaims;
-import com.ll.commars.domain.auth.token.entity.TokenValue;
 import com.ll.commars.domain.auth.token.entity.RefreshToken;
+import com.ll.commars.domain.auth.token.entity.TokenValue;
 import com.ll.commars.domain.member.entity.Member;
 import com.ll.commars.domain.member.repository.MemberRepository;
 import com.ll.commars.global.exception.CustomException;
@@ -32,7 +27,6 @@ public class AuthService {
 
 	private final TokenProvider tokenProvider;
 	private final MemberRepository memberRepository;
-	private final AuthenticationManager authenticationManager;
 	private final RedisTemplate<String, String> redisTemplate;
 
 	public void signOut(TokenValue accessTokenValue) {
@@ -76,27 +70,6 @@ public class AuthService {
 			newAccessToken.token().value(),
 			newRefreshToken.token().value()
 		);
-	}
-
-	public SignInResponse signIn(SignInRequest request) {
-		Authentication authentication = authenticationManager.authenticate(
-			new UsernamePasswordAuthenticationToken(request.email(), request.password())
-		);
-
-		String email = authentication.getName();
-		Member member = memberRepository.findByEmail(email)
-			.orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
-
-		AccessToken accessToken = tokenProvider.generateAccessToken(member);
-		RefreshToken refreshToken = tokenProvider.generateRefreshToken(member);
-
-		redisTemplate.opsForValue().set(
-			"refreshToken" + member.getId(),
-			refreshToken.token().value(),
-			Duration.ofMillis(refreshToken.expiration())
-		);
-
-		return new SignInResponse(accessToken.token().value(), refreshToken.token().value());
 	}
 
 	@Transactional
